--- app/src/main/java/io/legado/app/help/http/OkHttpExtensions.kt
@@ -1,49 +1,0 @@
-package io.legado.app.help.http
-
-import io.legado.app.utils.EncodingDetect
-import io.legado.app.utils.UTF8BOMFighter
-import kotlinx.coroutines.suspendCancellableCoroutine
-import okhttp3.Call
-import okhttp3.Callback
-import okhttp3.Response
-import okhttp3.ResponseBody
-import java.io.IOException
-import java.nio.charset.Charset
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-
-suspend fun Call.await(): Response = suspendCancellableCoroutine { block ->
-
-    block.invokeOnCancellation {
-        cancel()
-    }
-
-    enqueue(object : Callback {
-        override fun onFailure(call: Call, e: IOException) {
-            block.resumeWithException(e)
-        }
-
-        override fun onResponse(call: Call, response: Response) {
-            block.resume(response)
-        }
-    })
-
-}
-
-fun ResponseBody.text(encode: String? = null): String {
-    val responseBytes = UTF8BOMFighter.removeUTF8BOM(bytes())
-    var charsetName: String? = encode
-
-    charsetName?.let {
-        return String(responseBytes, Charset.forName(charsetName))
-    }
-
-    //根据http头判断
-    contentType()?.charset()?.let {
-        return String(responseBytes, it)
-    }
-
-    //根据内容判断
-    charsetName = EncodingDetect.getHtmlEncode(responseBytes)
-    return String(responseBytes, Charset.forName(charsetName))
-}
