+++ app/src/main/java/io/legado/app/lib/webdav/WebDav.kt
@@ -1,7 +1,7 @@
 package io.legado.app.lib.webdav
 
-import io.legado.app.help.http.HttpHelper
 import io.legado.app.help.http.await
+import io.legado.app.help.http.okHttpClient
 import okhttp3.*
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.RequestBody.Companion.asRequestBody
@@ -113,7 +113,7 @@ class WebDav(urlStr: String) {
                 request.header("Authorization", Credentials.basic(it.user, it.pass))
             }
             request.header("Depth", if (depth < 0) "infinity" else depth.toString())
-            return HttpHelper.client.newCall(request.build()).await()
+            return okHttpClient.newCall(request.build()).await()
         }
         return null
     }
@@ -227,7 +227,7 @@ class WebDav(urlStr: String) {
         HttpAuth.auth?.let {
             requestBuilder.header("Authorization", Credentials.basic(it.user, it.pass))
         }
-        val response = HttpHelper.client.newCall(requestBuilder.build()).await()
+        val response = okHttpClient.newCall(requestBuilder.build()).await()
         return response.isSuccessful
     }
 
@@ -239,7 +239,7 @@ class WebDav(urlStr: String) {
             val requestBuilder = Request.Builder()
                 .url(url)
                 .addHeader("Authorization", Credentials.basic(auth.user, auth.pass))
-            return HttpHelper.client.newCall(request = requestBuilder.build()).await()
+            return okHttpClient.newCall(request = requestBuilder.build()).await()
                 .body?.byteStream()
         }
         return null
